new version

+++ project_api/src/main/java/com/example/project_api/models/repository/ClubRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ClubRepository extends JpaRepository<Club, Integer> {
+public interface ClubRepository extends JpaRepository<Club, Long> {
    Optional<Club> findByTitle(String title);
    Optional<Club> findByUserId(long userId);
